made changes to tests, ignore one exception that can be tested, added assert to deletion methods, added test file for models
